backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maillens-api"
version = "0.1.0"
description = "MailLens FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.60.0",
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "python-dotenv==1.0.1",
  "sqlalchemy==2.0.34",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "pydantic==2.8.2",
  "pydantic-settings==2.5.2",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "httpx==0.27.2",
  "pinecone==5.0.1", 
  "tenacity==8.5.0" ,
  "celery[redis]==5.4.0", # ← add
  "redis==5.0.7",
  "google-api-python-client>=2.0.0" ,
  "tiktoken>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]           # only package our FastAPI app
exclude = [
  "alembic*", "infra*", "worker*", "tests*", "myenv*", ".venv*", "venv*"
]

[project.optional-dependencies]
dev = ["ipython", "ruff", "black"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]